Optimized Text class loadtextImg function It now caches the data so it only has to caclulate the data once per font also fixed a time glitch with the save data

+++ idleFisher/saveData.h
@@ -495,8 +495,8 @@ struct Fdata {
 
 struct FsaveData {
     // stored in milliseconds
-    double startTime;
-    double lastPlayed;
+    long long startTime;
+    long long lastPlayed;
 
     // player loc
     vector playerLoc = { 500, 500 };
